Zd.3.5 modyfikacja

+++ src/main/java/com/kodilla/basic_assertion/ResultChecker.java
@@ -1,7 +1,7 @@
 package com.kodilla.basic_assertion;
 
 public class ResultChecker {
-    public static boolean assertEquals(int expected, int actual) {
+    public static boolean assertEquals(double expected, int actual) {
         return expected == actual;
     }
